Replaced flask templating with jquery requests.
The objective is to take the pages currently served by this flask application
and host them on github pages (which server only static html).
In order to do this I'm replacing the flask templating with jquery calls
and making the flask application API only.
The flask app can then be hosted elsewhere and used as a microservice.