Created materialized views for expensive queries. Modified API to use new views.
The getNetwork and getGsmSignal functions from the API performed multiple expensive
database queries and processing.
Each function has been rewritten as single SQL queries, while the new queries
may be more expensive than the previous functions their results can now be
stored as materialized views in the database which means they only have to be performed
once a month when a new report is imported and the data changes.